+++ developer_scripts/update_container_tags.sh
@@ -9,8 +9,16 @@ if [[ "$target" == "main" ]]; then
   search_pattern='container: "ghcr\.io/stjudecloud/([^:]*):branch-[^-]*-(.*)"'
   replacement='container: "ghcr.io/stjudecloud/\1:\2"'
 else
-  search_pattern='container: "ghcr\.io/stjudecloud/([^:]*):(.*)"'
-  replacement='container: "ghcr.io/stjudecloud/\1:branch-'$target'-\2"'
+  search_pattern='container: "ghcr\.io/stjudecloud/([^:]*):(branch-[^-]*-)?(.*)"'
+  replacement='container: "ghcr.io/stjudecloud/\1:branch-'$target'-\3"'
 fi
 
-find . -not -path '.git' -a -type f -name '*.wdl' | xargs sed -Eri '' "s,$search_pattern,$replacement,g"
+# freeBSD sed (including the Mac version) requires a space and then an argument to -i
+if [ "$(uname)" == "Darwin" ]
+then
+  inplace_arg='-i ""'
+else
+  inplace_arg='-i'
+fi
+
+find . -not -path '.git' -a -type f -name '*.wdl' | xargs sed -Er $inplace_arg "s,$search_pattern,$replacement,g"